Upgrade specs and benchmarks to .NET 6.0 (#142)
* Update to .NET 6.0
Replace Obsolete BenchmarkDot APIs with preferred approaches.

* Revert unnecessary upgrade

The Ais.Net project itself does not benefit at all from being upgraded to .NET 6.0. It turns out that all the gains can be had simply by loading the existing library into a .NET 6.0 application. The existing library targetting .netstandard2.1 performs exactly as well as one targetting .NET 6.0 directly (even if we upgrade to the latest System.Io.Pipelines) so there's no need for us to do a new release.

* Remove SourceLink from specs & benchmark

We don't publish the spec or benchmark projects as a library so there's no reason for them to use sourcelink.

* Add .NET 6.0 SDK to build

* Remove .NET Core 2.1 support

Looks like it won't build unless we do this.

* I hate yaml

* Attempt to fix broken build

Co-authored-by: Howard van Rooijen <Howard.vanRooijen@endjin.com>